+++ documentation/allclasses-index.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
 <title>All Classes</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
+<meta name="dc.created" content="2020-07-21">
 <meta name="description" content="class index">
 <meta name="generator" content="javadoc/AllClassesIndexWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -79,22 +79,28 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </thead>
 <tbody>
 <tr class="altColor" id="i0">
-<td class="colFirst"><a href="br/com/fiap/Chat.html" title="class in br.com.fiap">Chat</a></td>
+<td class="colFirst"><a href="br/com/fiap/App.html" title="class in br.com.fiap">App</a></td>
 <th class="colLast" scope="row">&nbsp;</th>
 </tr>
 <tr class="rowColor" id="i1">
-<td class="colFirst"><a href="br/com/fiap/Helper.html" title="class in br.com.fiap">Helper</a></td>
-<th class="colLast" scope="row">&nbsp;</th>
+<td class="colFirst"><a href="br/com/fiap/AppTest.html" title="class in br.com.fiap">AppTest</a></td>
+<th class="colLast" scope="row">
+<div class="block">Unit test for simple App.</div>
+</th>
 </tr>
 <tr class="altColor" id="i2">
-<td class="colFirst"><a href="br/com/fiap/Main.html" title="class in br.com.fiap">Main</a></td>
+<td class="colFirst"><a href="br/com/fiap/Chat.html" title="class in br.com.fiap">Chat</a></td>
 <th class="colLast" scope="row">&nbsp;</th>
 </tr>
 <tr class="rowColor" id="i3">
-<td class="colFirst"><a href="br/com/fiap/Match.html" title="class in br.com.fiap">Match</a></td>
+<td class="colFirst"><a href="br/com/fiap/Helper.html" title="class in br.com.fiap">Helper</a></td>
 <th class="colLast" scope="row">&nbsp;</th>
 </tr>
 <tr class="altColor" id="i4">
+<td class="colFirst"><a href="br/com/fiap/Match.html" title="class in br.com.fiap">Match</a></td>
+<th class="colLast" scope="row">&nbsp;</th>
+</tr>
+<tr class="rowColor" id="i5">
 <td class="colFirst"><a href="br/com/fiap/Message.html" title="class in br.com.fiap">Message</a></td>
 <th class="colLast" scope="row">&nbsp;</th>
 </tr>
