+++ documentation/br/com/fiap/AppTest.html
@@ -3,10 +3,10 @@
 <html lang="pt">
 <head>
 <!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
-<title>Match</title>
+<title>AppTest</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-07-21">
-<meta name="description" content="declaration: package: br.com.fiap, class: Match">
+<meta name="description" content="declaration: package: br.com.fiap, class: AppTest">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.css" title="Style">
@@ -20,7 +20,7 @@
 <script type="text/javascript" src="../../../script-dir/jquery-ui.js"></script>
 </head>
 <body class="class-declaration">
-<script type="text/javascript">var data = {"i0":10,"i1":10,"i2":10,"i3":10};
+<script type="text/javascript">var data = {"i0":10};
 var tabs = {65535:["t0","All Methods"],2:["t2","Instance Methods"],8:["t4","Concrete Methods"]};
 var altColor = "altColor";
 var rowColor = "rowColor";
@@ -88,16 +88,17 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <main role="main">
 <div class="header">
 <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="package-summary.html">br.com.fiap</a></div>
-<h1 title="Class Match" class="title">Class Match</h1>
+<h1 title="Class AppTest" class="title">Class AppTest</h1>
 </div>
 <div class="contentContainer">
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">br.com.fiap.Match</div>
+<div class="inheritance">br.com.fiap.AppTest</div>
 </div>
 <section class="description">
 <hr>
-<pre>public class <span class="typeNameLabel">Match</span>
+<pre>public class <span class="typeNameLabel">AppTest</span>
 extends java.lang.Object</pre>
+<div class="block">Unit test for simple App.</div>
 </section>
 <section class="summary">
 <ul class="blockList">
@@ -118,11 +119,8 @@ extends java.lang.Object</pre>
 </thead>
 <tbody>
 <tr class="altColor">
-<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(java.lang.String,int)">Match</a></span>&#8203;(java.lang.String&nbsp;value,
-int&nbsp;distance)</code></th>
-<td class="colLast">
-<div class="block">Método que instancia um novo objeto do tipo Match</div>
-</td>
+<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E()">AppTest</a></span>()</code></th>
+<td class="colLast">&nbsp;</td>
 </tr>
 </tbody>
 </table>
@@ -148,31 +146,10 @@ int&nbsp;distance)</code></th>
 </thead>
 <tbody>
 <tr class="altColor" id="i0">
-<td class="colFirst"><code>int</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#getDistance()">getDistance</a></span>()</code></th>
-<td class="colLast">
-<div class="block">Método que obtém a distância</div>
-</td>
-</tr>
-<tr class="rowColor" id="i1">
-<td class="colFirst"><code>java.lang.String</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#getValue()">getValue</a></span>()</code></th>
-<td class="colLast">
-<div class="block">Método que obtém o valor da mensagem</div>
-</td>
-</tr>
-<tr class="altColor" id="i2">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#setDistance(int)">setDistance</a></span>&#8203;(int&nbsp;distance)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#shouldAnswerWithTrue()">shouldAnswerWithTrue</a></span>()</code></th>
 <td class="colLast">
-<div class="block">Método que atribui o valor a distancia</div>
-</td>
-</tr>
-<tr class="rowColor" id="i3">
-<td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#setValue(java.lang.String)">setValue</a></span>&#8203;(java.lang.String&nbsp;value)</code></th>
-<td class="colLast">
-<div class="block">Método que atribui o valor a String value</div>
+<div class="block">Rigorous Test :-)</div>
 </td>
 </tr>
 </tbody>
@@ -199,15 +176,8 @@ int&nbsp;distance)</code></th>
 <ul class="blockList">
 <li class="blockList">
 <section class="detail">
-<h3><a id="&lt;init&gt;(java.lang.String,int)">Match</a></h3>
-<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="memberName">Match</span>&#8203;(<span class="arguments">java.lang.String&nbsp;value,
-int&nbsp;distance)</span></div>
-<div class="block">Método que instancia um novo objeto do tipo Match</div>
-<dl>
-<dt><span class="paramLabel">Parameters:</span></dt>
-<dd><code>value</code> - String que se deseja obter a distância</dd>
-<dd><code>distance</code> - inteiro com a distância</dd>
-</dl>
+<h3><a id="&lt;init&gt;()">AppTest</a></h3>
+<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="memberName">AppTest</span>()</div>
 </section>
 </li>
 </ul>
@@ -222,46 +192,9 @@ int&nbsp;distance)</span></div>
 <ul class="blockList">
 <li class="blockList">
 <section class="detail">
-<h3><a id="getValue()">getValue</a></h3>
-<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="returnType">java.lang.String</span>&nbsp;<span class="memberName">getValue</span>()</div>
-<div class="block">Método que obtém o valor da mensagem</div>
-<dl>
-<dt><span class="returnLabel">Returns:</span></dt>
-<dd>retorna valor obtido</dd>
-</dl>
-</section>
-</li>
-<li class="blockList">
-<section class="detail">
-<h3><a id="setValue(java.lang.String)">setValue</a></h3>
-<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="returnType">void</span>&nbsp;<span class="memberName">setValue</span>&#8203;(<span class="arguments">java.lang.String&nbsp;value)</span></div>
-<div class="block">Método que atribui o valor a String value</div>
-<dl>
-<dt><span class="paramLabel">Parameters:</span></dt>
-<dd><code>value</code> - do tipo String</dd>
-</dl>
-</section>
-</li>
-<li class="blockList">
-<section class="detail">
-<h3><a id="getDistance()">getDistance</a></h3>
-<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="returnType">int</span>&nbsp;<span class="memberName">getDistance</span>()</div>
-<div class="block">Método que obtém a distância</div>
-<dl>
-<dt><span class="returnLabel">Returns:</span></dt>
-<dd>retorna a distância obtida</dd>
-</dl>
-</section>
-</li>
-<li class="blockList">
-<section class="detail">
-<h3><a id="setDistance(int)">setDistance</a></h3>
-<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="returnType">void</span>&nbsp;<span class="memberName">setDistance</span>&#8203;(<span class="arguments">int&nbsp;distance)</span></div>
-<div class="block">Método que atribui o valor a distancia</div>
-<dl>
-<dt><span class="paramLabel">Parameters:</span></dt>
-<dd><code>distance</code> - do tipo inteiro</dd>
-</dl>
+<h3><a id="shouldAnswerWithTrue()">shouldAnswerWithTrue</a></h3>
+<div class="memberSignature"><span class="modifiers">public</span>&nbsp;<span class="returnType">void</span>&nbsp;<span class="memberName">shouldAnswerWithTrue</span>()</div>
+<div class="block">Rigorous Test :-)</div>
 </section>
 </li>
 </ul>
