+++ documentation/br/com/fiap/package-summary.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
 <title>br.com.fiap</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
+<meta name="dc.created" content="2020-07-21">
 <meta name="description" content="declaration: package: br.com.fiap">
 <meta name="generator" content="javadoc/PackageWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
@@ -82,6 +82,16 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </thead>
 <tbody>
 <tr class="altColor">
+<th class="colFirst" scope="row"><a href="App.html" title="class in br.com.fiap">App</a></th>
+<td class="colLast">&nbsp;</td>
+</tr>
+<tr class="rowColor">
+<th class="colFirst" scope="row"><a href="AppTest.html" title="class in br.com.fiap">AppTest</a></th>
+<td class="colLast">
+<div class="block">Unit test for simple App.</div>
+</td>
+</tr>
+<tr class="altColor">
 <th class="colFirst" scope="row"><a href="Chat.html" title="class in br.com.fiap">Chat</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
@@ -90,14 +100,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="Main.html" title="class in br.com.fiap">Main</a></th>
-<td class="colLast">&nbsp;</td>
-</tr>
-<tr class="rowColor">
 <th class="colFirst" scope="row"><a href="Match.html" title="class in br.com.fiap">Match</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
-<tr class="altColor">
+<tr class="rowColor">
 <th class="colFirst" scope="row"><a href="Message.html" title="class in br.com.fiap">Message</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
