+++ documentation/br/com/fiap/package-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
 <title>br.com.fiap Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
+<meta name="dc.created" content="2020-07-21">
 <meta name="description" content="tree: package: br.com.fiap">
 <meta name="generator" content="javadoc/PackageTreeWriter">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
@@ -73,9 +73,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
+<li class="circle">br.com.fiap.<a href="App.html" title="class in br.com.fiap"><span class="typeNameLink">App</span></a></li>
+<li class="circle">br.com.fiap.<a href="AppTest.html" title="class in br.com.fiap"><span class="typeNameLink">AppTest</span></a></li>
 <li class="circle">br.com.fiap.<a href="Chat.html" title="class in br.com.fiap"><span class="typeNameLink">Chat</span></a></li>
 <li class="circle">br.com.fiap.<a href="Helper.html" title="class in br.com.fiap"><span class="typeNameLink">Helper</span></a></li>
-<li class="circle">br.com.fiap.<a href="Main.html" title="class in br.com.fiap"><span class="typeNameLink">Main</span></a></li>
 <li class="circle">br.com.fiap.<a href="Match.html" title="class in br.com.fiap"><span class="typeNameLink">Match</span></a></li>
 <li class="circle">br.com.fiap.<a href="Message.html" title="class in br.com.fiap"><span class="typeNameLink">Message</span></a></li>
 </ul>
