+++ documentation/help-doc.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
 <title>API Help</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
+<meta name="dc.created" content="2020-07-21">
 <meta name="description" content="help">
 <meta name="generator" content="javadoc/HelpWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
