+++ documentation/index-files/index-10.html
@@ -3,10 +3,10 @@
 <html lang="pt">
 <head>
 <!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
-<title>A-Index</title>
+<title>S-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-07-21">
-<meta name="description" content="index: A">
+<meta name="description" content="index: S">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.css" title="Style">
@@ -67,21 +67,31 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <div class="header">
 <h1>Index</h1>
 </div>
-<div class="contentContainer"><a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:A">
+<div class="contentContainer"><a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:S">
 <!--   -->
 </a>
-<h2 class="title">A</h2>
+<h2 class="title">S</h2>
 <dl>
-<dt><a href="../br/com/fiap/App.html" title="class in br.com.fiap"><span class="typeNameLink">App</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
-<dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/App.html#%3Cinit%3E()">App()</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/App.html" title="class in br.com.fiap">App</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../br/com/fiap/AppTest.html" title="class in br.com.fiap"><span class="typeNameLink">AppTest</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Match.html#setDistance(int)">setDistance(int)</a></span> - Method in class br.com.fiap.<a href="../br/com/fiap/Match.html" title="class in br.com.fiap">Match</a></dt>
 <dd>
-<div class="block">Unit test for simple App.</div>
+<div class="block">Método que atribui o valor a distancia</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/AppTest.html#%3Cinit%3E()">AppTest()</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/AppTest.html" title="class in br.com.fiap">AppTest</a></dt>
-<dd>&nbsp;</dd>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Message.html#setStage(int)">setStage(int)</a></span> - Method in class br.com.fiap.<a href="../br/com/fiap/Message.html" title="class in br.com.fiap">Message</a></dt>
+<dd>
+<div class="block">Método para setar o estágio</div>
+</dd>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Match.html#setValue(java.lang.String)">setValue(String)</a></span> - Method in class br.com.fiap.<a href="../br/com/fiap/Match.html" title="class in br.com.fiap">Match</a></dt>
+<dd>
+<div class="block">Método que atribui o valor a String value</div>
+</dd>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Message.html#setValue(java.lang.String)">setValue(String)</a></span> - Method in class br.com.fiap.<a href="../br/com/fiap/Message.html" title="class in br.com.fiap">Message</a></dt>
+<dd>
+<div class="block">Método para setar o valor da mensagem</div>
+</dd>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/AppTest.html#shouldAnswerWithTrue()">shouldAnswerWithTrue()</a></span> - Method in class br.com.fiap.<a href="../br/com/fiap/AppTest.html" title="class in br.com.fiap">AppTest</a></dt>
+<dd>
+<div class="block">Rigorous Test :-)</div>
+</dd>
 </dl>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
 </main>
