+++ documentation/index-files/index-6.html
@@ -2,11 +2,11 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
-<title>I-Index</title>
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
+<title>H-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
-<meta name="description" content="index: I">
+<meta name="dc.created" content="2020-07-21">
+<meta name="description" content="index: H">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.css" title="Style">
@@ -67,17 +67,17 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <div class="header">
 <h1>Index</h1>
 </div>
-<div class="contentContainer"><a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">H</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:I">
+<div class="contentContainer"><a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:H">
 <!--   -->
 </a>
-<h2 class="title">I</h2>
+<h2 class="title">H</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/Chat.html#init()">init()</a></span> - Static method in class br.com.fiap.<a href="../br/com/fiap/Chat.html" title="class in br.com.fiap">Chat</a></dt>
-<dd>
-<div class="block">Método que inicializa a base</div>
-</dd>
+<dt><a href="../br/com/fiap/Helper.html" title="class in br.com.fiap"><span class="typeNameLink">Helper</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
+<dd>&nbsp;</dd>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Helper.html#%3Cinit%3E()">Helper()</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/Helper.html" title="class in br.com.fiap">Helper</a></dt>
+<dd>&nbsp;</dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">H</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
 </main>
 <footer role="contentinfo">
 <nav role="navigation">
