+++ documentation/index-files/index-8.html
@@ -2,11 +2,11 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
-<title>M-Index</title>
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
+<title>L-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
-<meta name="description" content="index: M">
+<meta name="dc.created" content="2020-07-21">
+<meta name="description" content="index: L">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.css" title="Style">
@@ -67,33 +67,17 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <div class="header">
 <h1>Index</h1>
 </div>
-<div class="contentContainer"><a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">H</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:M">
+<div class="contentContainer"><a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a><a id="I:L">
 <!--   -->
 </a>
-<h2 class="title">M</h2>
+<h2 class="title">L</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/Main.html#main(java.lang.String%5B%5D)">main(String[])</a></span> - Static method in class br.com.fiap.<a href="../br/com/fiap/Main.html" title="class in br.com.fiap">Main</a></dt>
+<dt><span class="memberNameLink"><a href="../br/com/fiap/Chat.html#loadFaq(java.io.File)">loadFaq(File)</a></span> - Static method in class br.com.fiap.<a href="../br/com/fiap/Chat.html" title="class in br.com.fiap">Chat</a></dt>
 <dd>
-<div class="block">Método responsável por controlar o fluxo da conversa e receber as mensagens do usuário</div>
-</dd>
-<dt><a href="../br/com/fiap/Main.html" title="class in br.com.fiap"><span class="typeNameLink">Main</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
-<dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/Main.html#%3Cinit%3E()">Main()</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/Main.html" title="class in br.com.fiap">Main</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../br/com/fiap/Match.html" title="class in br.com.fiap"><span class="typeNameLink">Match</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
-<dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/Match.html#%3Cinit%3E(java.lang.String,int)">Match(String, int)</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/Match.html" title="class in br.com.fiap">Match</a></dt>
-<dd>
-<div class="block">Método que instancia um novo objeto do tipo Match</div>
-</dd>
-<dt><a href="../br/com/fiap/Message.html" title="class in br.com.fiap"><span class="typeNameLink">Message</span></a> - Class in <a href="../br/com/fiap/package-summary.html">br.com.fiap</a></dt>
-<dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../br/com/fiap/Message.html#%3Cinit%3E(int,java.lang.String)">Message(int, String)</a></span> - Constructor for class br.com.fiap.<a href="../br/com/fiap/Message.html" title="class in br.com.fiap">Message</a></dt>
-<dd>
-<div class="block">Método que instancia um novo objeto do tipo Message</div>
+<div class="block">Método responsável por ler e carregar os arquivos .csv</div>
 </dd>
 </dl>
-<a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">H</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
+<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">B</a>&nbsp;<a href="index-3.html">C</a>&nbsp;<a href="index-4.html">D</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
 </main>
 <footer role="contentinfo">
 <nav role="navigation">
