+++ documentation/member-search-index.js
@@ -1,1 +1,1 @@
-memberSearchIndex = [{"p":"br.com.fiap","c":"Chat","l":"Chat()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Helper","l":"distance(String, String)","url":"distance(java.lang.String,java.lang.String)"},{"p":"br.com.fiap","c":"Chat","l":"getAnswer(Message)","url":"getAnswer(br.com.fiap.Message)"},{"p":"br.com.fiap","c":"Match","l":"getDistance()"},{"p":"br.com.fiap","c":"Chat","l":"getFaq(String)","url":"getFaq(java.lang.String)"},{"p":"br.com.fiap","c":"Helper","l":"getMatch(String, HashMap<String, String>)","url":"getMatch(java.lang.String,java.util.HashMap)"},{"p":"br.com.fiap","c":"Message","l":"getStage()"},{"p":"br.com.fiap","c":"Match","l":"getValue()"},{"p":"br.com.fiap","c":"Message","l":"getValue()"},{"p":"br.com.fiap","c":"Message","l":"goForwardStage()"},{"p":"br.com.fiap","c":"Helper","l":"Helper()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Chat","l":"init()"},{"p":"br.com.fiap","c":"Chat","l":"loadFaq(File)","url":"loadFaq(java.io.File)"},{"p":"br.com.fiap","c":"Main","l":"Main()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Main","l":"main(String[])","url":"main(java.lang.String[])"},{"p":"br.com.fiap","c":"Match","l":"Match(String, int)","url":"%3Cinit%3E(java.lang.String,int)"},{"p":"br.com.fiap","c":"Message","l":"Message(int, String)","url":"%3Cinit%3E(int,java.lang.String)"},{"p":"br.com.fiap","c":"Match","l":"setDistance(int)"},{"p":"br.com.fiap","c":"Message","l":"setStage(int)"},{"p":"br.com.fiap","c":"Match","l":"setValue(String)","url":"setValue(java.lang.String)"},{"p":"br.com.fiap","c":"Message","l":"setValue(String)","url":"setValue(java.lang.String)"}]
+memberSearchIndex = [{"p":"br.com.fiap","c":"App","l":"App()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"AppTest","l":"AppTest()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Chat","l":"Chat()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Helper","l":"distance(String, String)","url":"distance(java.lang.String,java.lang.String)"},{"p":"br.com.fiap","c":"Chat","l":"getAnswer(Message)","url":"getAnswer(br.com.fiap.Message)"},{"p":"br.com.fiap","c":"Match","l":"getDistance()"},{"p":"br.com.fiap","c":"Chat","l":"getFaq(String)","url":"getFaq(java.lang.String)"},{"p":"br.com.fiap","c":"Helper","l":"getMatch(String, HashMap<String, String>)","url":"getMatch(java.lang.String,java.util.HashMap)"},{"p":"br.com.fiap","c":"Message","l":"getStage()"},{"p":"br.com.fiap","c":"Match","l":"getValue()"},{"p":"br.com.fiap","c":"Message","l":"getValue()"},{"p":"br.com.fiap","c":"Message","l":"goForwardStage()"},{"p":"br.com.fiap","c":"Helper","l":"Helper()","url":"%3Cinit%3E()"},{"p":"br.com.fiap","c":"Chat","l":"init()"},{"p":"br.com.fiap","c":"Chat","l":"loadFaq(File)","url":"loadFaq(java.io.File)"},{"p":"br.com.fiap","c":"App","l":"main(String[])","url":"main(java.lang.String[])"},{"p":"br.com.fiap","c":"Match","l":"Match(String, int)","url":"%3Cinit%3E(java.lang.String,int)"},{"p":"br.com.fiap","c":"Message","l":"Message(int, String)","url":"%3Cinit%3E(int,java.lang.String)"},{"p":"br.com.fiap","c":"Match","l":"setDistance(int)"},{"p":"br.com.fiap","c":"Message","l":"setStage(int)"},{"p":"br.com.fiap","c":"Match","l":"setValue(String)","url":"setValue(java.lang.String)"},{"p":"br.com.fiap","c":"Message","l":"setValue(String)","url":"setValue(java.lang.String)"},{"p":"br.com.fiap","c":"AppTest","l":"shouldAnswerWithTrue()"}]
