+++ documentation/overview-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="pt">
 <head>
-<!-- Generated by javadoc (13.0.2) on Sun Jul 19 12:16:08 BRT 2020 -->
+<!-- Generated by javadoc (13.0.2) on Tue Jul 21 17:57:20 BRT 2020 -->
 <title>Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2020-07-19">
+<meta name="dc.created" content="2020-07-21">
 <meta name="description" content="class tree">
 <meta name="generator" content="javadoc/TreeWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -77,9 +77,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
+<li class="circle">br.com.fiap.<a href="br/com/fiap/App.html" title="class in br.com.fiap"><span class="typeNameLink">App</span></a></li>
+<li class="circle">br.com.fiap.<a href="br/com/fiap/AppTest.html" title="class in br.com.fiap"><span class="typeNameLink">AppTest</span></a></li>
 <li class="circle">br.com.fiap.<a href="br/com/fiap/Chat.html" title="class in br.com.fiap"><span class="typeNameLink">Chat</span></a></li>
 <li class="circle">br.com.fiap.<a href="br/com/fiap/Helper.html" title="class in br.com.fiap"><span class="typeNameLink">Helper</span></a></li>
-<li class="circle">br.com.fiap.<a href="br/com/fiap/Main.html" title="class in br.com.fiap"><span class="typeNameLink">Main</span></a></li>
 <li class="circle">br.com.fiap.<a href="br/com/fiap/Match.html" title="class in br.com.fiap"><span class="typeNameLink">Match</span></a></li>
 <li class="circle">br.com.fiap.<a href="br/com/fiap/Message.html" title="class in br.com.fiap"><span class="typeNameLink">Message</span></a></li>
 </ul>
