+++ documentation/type-search-index.js
@@ -1,1 +1,1 @@
-typeSearchIndex = [{"l":"All Classes","url":"allclasses-index.html"},{"p":"br.com.fiap","l":"Chat"},{"p":"br.com.fiap","l":"Helper"},{"p":"br.com.fiap","l":"Main"},{"p":"br.com.fiap","l":"Match"},{"p":"br.com.fiap","l":"Message"}]
+typeSearchIndex = [{"l":"All Classes","url":"allclasses-index.html"},{"p":"br.com.fiap","l":"App"},{"p":"br.com.fiap","l":"AppTest"},{"p":"br.com.fiap","l":"Chat"},{"p":"br.com.fiap","l":"Helper"},{"p":"br.com.fiap","l":"Match"},{"p":"br.com.fiap","l":"Message"}]
