--- src/br/com/fiap/Main.java
@@ -1,77 +1,0 @@
-package br.com.fiap;
-
-import java.util.List;
-
-import com.pengrad.telegrambot.TelegramBot;
-import com.pengrad.telegrambot.TelegramBotAdapter;
-import com.pengrad.telegrambot.model.Update;
-import com.pengrad.telegrambot.model.request.ChatAction;
-import com.pengrad.telegrambot.request.GetUpdates;
-import com.pengrad.telegrambot.request.SendChatAction;
-import com.pengrad.telegrambot.request.SendMessage;
-import com.pengrad.telegrambot.response.BaseResponse;
-import com.pengrad.telegrambot.response.GetUpdatesResponse;
-import com.pengrad.telegrambot.response.SendResponse;
-
-public class Main {
-
-	/**
-	 * Método responsável por controlar o fluxo da conversa e receber as mensagens do usuário
-	 * @param args array de String
-	 * @throws Exception caso ocorra alguma exceção em tempo de execução
-	 */
-	public static void main(String[] args) throws Exception{
-		
-		Chat.init();
-
-		//Criação do objeto bot com as informações de acesso
-		TelegramBot bot = TelegramBotAdapter.build("1250519022:AAHkV4tRjvpnqRuXTkzJDeP2M7mg5teheAw");
-
-		//objeto responsável por receber as mensagens
-		GetUpdatesResponse updatesResponse;
-		//objeto responsável por gerenciar o envio de respostas
-		SendResponse sendResponse;
-		//objeto responsável por gerenciar o envio de ações do chat
-		BaseResponse baseResponse;
-		
-		//controle de off-set, isto é, a partir deste ID será lido as mensagens pendentes na fila
-		int m=0;
-		
-		Message message = new Message(0, "");
-		
-		//loop infinito pode ser alterado por algum timer de intervalo curto
-		while (true){
-		
-			//executa comando no Telegram para obter as mensagens pendentes a partir de um off-set (limite inicial)
-			updatesResponse =  bot.execute(new GetUpdates().limit(100).offset(m));
-			
-			//lista de mensagens
-			List<Update> updates = updatesResponse.updates();
-
-			//análise de cada ação da mensagem
-			for (Update update : updates) {
-				
-				//atualização do off-set
-				m = update.updateId()+1;
-				
-				message.setValue(update.message().text());
-				
-				System.out.println("Recebendo mensagem: "+ message.getValue());
-				
-				//envio de "Escrevendo" antes de enviar a resposta
-				baseResponse = bot.execute(new SendChatAction(update.message().chat().id(), ChatAction.typing.name()));
-				//verificação de ação de chat foi enviada com sucesso
-				System.out.println("Resposta de Chat Action Enviada? " + baseResponse.isOk());
-				
-				//envio da mensagem de resposta
-				sendResponse = bot.execute(new SendMessage(update.message().chat().id(), Chat.getAnswer(message)));
-				//verificação de mensagem enviada com sucesso
-				System.out.println("Mensagem Enviada? " +sendResponse.isOk());
-				
-			}
-
-		}
-
-	}
-
-}
