+++ src/main/java/br/com/fiap/Helper.java
@@ -46,7 +46,7 @@ public class Helper {
 				chave = key;
 			}
 		};
-		Match m = new Match(chave, menor);
+        Match m = new Match(chave, menor);
 		return m;
 	}
 }
